+++ src/MyLab.Search.Indexer/Services/IIndexResourceProvider.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using MyLab.Search.EsAdapter;
+
+namespace MyLab.Search.Indexer.Services
+{
+    public interface IIndexResourceProvider
+    {
+        Task<string> ReadFileAsync(string idxId, string filename);
+    }
+}