+++ src/MyLab.Search.Indexer/Services/IPushIndexer.cs
@@ -4,13 +4,14 @@ using System.Threading;
 using System.Threading.Tasks;
 using MyLab.Log;
 using MyLab.Search.Indexer.DataContract;
+using MyLab.Search.Indexer.Options;
 using MyLab.Search.Indexer.Tools;
 
 namespace MyLab.Search.Indexer.Services
 {
     public interface IPushIndexer
     {
-        Task IndexAsync(string strEntity, string sourceId, NsOptions nsOptions, CancellationToken cancellationToken);
+        Task IndexAsync(string strEntity, string sourceId, IdxOptions idxOptions, CancellationToken cancellationToken);
     }
 
     class PushIndexer : IPushIndexer
@@ -22,15 +23,15 @@ namespace MyLab.Search.Indexer.Services
             _indexer = indexer;
         }
 
-        public Task IndexAsync(string strEntity, string sourceId, NsOptions nsOptions, CancellationToken cancellationToken)
+        public Task IndexAsync(string strEntity, string sourceId, IdxOptions idxOptions, CancellationToken cancellationToken)
         {
             if (sourceId == null) throw new ArgumentNullException(nameof(sourceId));
-            if (nsOptions == null) throw new ArgumentNullException(nameof(nsOptions));
+            if (idxOptions == null) throw new ArgumentNullException(nameof(idxOptions));
 
             if (string.IsNullOrWhiteSpace(strEntity))
                 throw new BadIndexingRequestException("Entity object is empty");
             
-            var sourceEntityDeserializer = new SourceEntityDeserializer(nsOptions.NewIndexStrategy == NewIndexStrategy.Auto);
+            var sourceEntityDeserializer = new SourceEntityDeserializer(idxOptions.NewIndexStrategy == NewIndexStrategy.Auto);
 
             DataSourceEntity entity;
 
@@ -50,15 +51,15 @@ namespace MyLab.Search.Indexer.Services
                     .AndFactIs("dump", TrimDump(strEntity))
                     .AndFactIs("source", sourceId);
 
-            if (entity.Properties.Keys.All(k => k != nsOptions.IdPropertyName))
+            if (entity.Properties.Keys.All(k => k != idxOptions.IdPropertyName))
                 throw new BadIndexingRequestException("Cant find ID property in the entity object")
                     .AndFactIs("dump", TrimDump(strEntity))
                     .AndFactIs("source", sourceId);
 
-            var preproc = new DsEntityPreprocessor(nsOptions);
+            var preproc = new DsEntityPreprocessor(idxOptions);
             var entForIndex = new[] { preproc.Process(entity) };
 
-            return _indexer.IndexAsync(nsOptions.NsId, entForIndex, cancellationToken);
+            return _indexer.IndexAsync(idxOptions.Id, entForIndex, cancellationToken);
         }
 
         string TrimDump(string dump)