+++ src/UnitTests/IndexerTaskLogicBehavior.stuff.cs
@@ -13,6 +13,7 @@ using MyLab.DbTest;
 using MyLab.Search.EsAdapter;
 using MyLab.Search.Indexer;
 using MyLab.Search.Indexer.DataContract;
+using MyLab.Search.Indexer.Options;
 using MyLab.Search.Indexer.Services;
 using MyLab.Search.Indexer.Tools;
 using MyLab.TaskApp;
@@ -42,7 +43,7 @@ namespace UnitTests
                 .AddSingleton<ISeedService, TestSeedService>()
                 .AddSingleton<IDataIndexer, TestIndexer>()
                 .AddSingleton<ITaskLogic, IndexerTaskLogic>()
-                .AddSingleton<INamespaceResourceProvider, NamespaceResourceProvider>()
+                .AddSingleton<IIndexResourceProvider, IndexResourceProvider>()
                 //.Configure(configureDbOptions)
                 .Configure<ElasticsearchOptions>(o =>
                 {